Transaction for WaitingList to ParticipationList
Implemented transaction when deleting character, if participation list is full, next person from WaitingList is put in Participaton list. Still have to figure out about  transaction isolation level and some edgecases.

+++ GMS/GMS - Data Access Layer/EventCharacterAccess.cs
@@ -60,13 +60,31 @@ namespace GMS___Data_Access_Layer
         {
             using (IDbConnection conn = GetConnection())
             {
-                int rowsAffected = conn.Execute(@"DELETE FROM EventCharacter WHERE eventID = @EventID AND characterName = @CharacterName", new { EventID = eventID, CharacterName = characterName });
+                int rowsAffected;
 
-                if (rowsAffected > 0)
+                if (isParticipantListFull(eventID))
                 {
-                    return true;
+                    //rowsAffected = conn.Execute(@"DELETE FROM EventCharacter WHERE eventID = @EventID AND characterName = @CharacterName", new { EventID = eventID, CharacterName = characterName });
+
+                    //if (rowsAffected > 0)
+                    //{
+                    //    return true;
+                    //}
+                    //return false;
+
+                    return MoveCharacterFromWaitingListToParticipantList(eventID, characterName);
+
                 }
-                return false;
+                else
+                {
+                    rowsAffected = conn.Execute(@"DELETE FROM EventCharacter WHERE eventID = @EventID AND characterName = @CharacterName", new { EventID = eventID, CharacterName = characterName });
+                    
+                    if (rowsAffected > 0)
+                    {
+                        return true;
+                    }
+                    return false;
+                }
             }
         }
 
@@ -78,6 +96,48 @@ namespace GMS___Data_Access_Layer
                 int signedUpCount = conn.ExecuteScalar<int>("SELECT COUNT(*) FROM EventCharacter WHERE eventID = " + eventID);
 
                 return signedUpCount == maxAmount;
+            }
+        }
+
+        public bool MoveCharacterFromWaitingListToParticipantList(int eventID, string characterName)
+        {
+            using (IDbConnection conn = GetConnection())
+            {
+
+                conn.Open();
+                using (var trans = conn.BeginTransaction())
+                {
+                    try
+                    {
+                        //Delete character from Participants list
+                        conn.Execute(@"DELETE FROM EventCharacter WHERE eventID = @EventID AND characterName = @CharacterName", new { EventID = eventID, CharacterName = characterName }, trans);
+
+                        //Select character from waiting list that signed up first
+                        IEnumerable<EventCharacter> selectedCharacter = conn.Query<EventCharacter>(@"SELECT TOP 1 * FROM EventCharacterWaitingList WHERE eventID = @EventID ORDER BY signUpDateTime ASC",
+                            new { EventID = eventID }, trans);
+
+                        //Delete that character from WaitingList
+                        conn.Execute(@"DELETE FROM EventCharacterWaitingList WHERE eventID = @EventID AND characterName = @CharacterName",
+                            selectedCharacter, trans);
+
+                        //Put that character in Participants List
+                        conn.Execute(@"INSERT INTO EventCharacter (eventID, characterName, characterRole, signUpDateTime) " +
+                                    "VALUES (@EventID, @CharacterName, @CharacterRole, @SignUpDateTime)", selectedCharacter, trans);
+
+                        trans.Commit();
+
+                        return true;
+                       
+                    }
+                    catch (Exception ex)
+                    {
+                        trans.Rollback();
+
+                        Console.WriteLine(ex.ToString()); // TODO change exception handling
+
+                        return false;
+                    }
+                }
             }
         }
     }
